requirements.txt
torch~=2.3.1
wandb~=0.17.1
PyYAML~=6.0.1
gymnasium~=0.29.1
tqdm~=4.66.4
click~=8.1.7